굼바

+++ Assets/Script/Enemy.cs
@@ -5,16 +5,19 @@ public class Enemy : MonoBehaviour
 {
  
     public object collider;
+    bool[] tags = new bool[2];
 
     void Start()
     {
         if(tag == "Traps")
         {
             collider = GetComponent<TilemapCollider2D>();
+            tags[0] = true;
         }
         else if(tag == "Enemy")
         {
             collider = GetComponent<Collider2D>();
+            tags[1] = true;
         }
     }
 
@@ -22,7 +25,15 @@ public class Enemy : MonoBehaviour
     {
         if(collision.gameObject.tag == "Player")
         {
-            SpawnManager.instance.Kill();
+            if (tags[0] == true)
+            {
+                SpawnManager.instance.Kill();
+            }
+            else if (tags[1] == true)
+            {
+                //SpawnManager.instance.Kill();
+            }
+            
         }
     }
     void Update()
